+++ tests/unit/test_create_project.py
@@ -11,7 +11,7 @@ unify = Client(auth)
 
 @responses.activate
 def test_create_project():
-    project_creation_spec = {
+    creation_spec = {
         "name": "Project 1",
         "description": "Mastering Project",
         "type": "DEDUP",
@@ -23,9 +23,9 @@ def test_create_project():
     projects_url = f"http://localhost:9100/api/versioned/v1/projects"
     project_url = f"http://localhost:9100/api/versioned/v1/projects/1"
 
-    responses.add(responses.POST, projects_url, json=project_creation_spec, status=204)
-    responses.add(responses.GET, project_url, json=project_creation_spec)
+    responses.add(responses.POST, projects_url, json=creation_spec, status=204)
+    responses.add(responses.GET, project_url, json=creation_spec)
 
-    u = unify.create_project(project_creation_spec)
+    u = unify.projects.create(creation_spec)
     p = unify.projects.by_resource_id("1")
     assert print(p) == print(u)