+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## 0.6.0-dev
+  **BREAKING CHANGES**
+  - [#150](https://github.com/Datatamer/unify-client-python/issues/150) Move `create_project` and `create_dataset` from client.py to corresponding collection.py
+
   **NEW FEATURES**
   - [#121](https://github.com/Datatamer/unify-client-python/issues/121) Fetches record clusters with data represented as a dataset.
 
@@ -18,7 +21,7 @@
   - [#114](https://github.com/Datatamer/unify-client-python/issues/114) Add support for generating pairs estimate
   - [#106](https://github.com/Datatamer/unify-client-python/issues/106) Add support for initializing a source dataset
   - [#107](https://github.com/Datatamer/unify-client-python/issues/107) Add support for creating a dataset attribute
-  
+
   **BUG FIXES**
   - [#118](https://github.com/Datatamer/unify-client-python/issues/118) Fix JSON sent for Dataset.update_records